mcp-practice1/requirements.txt
mcp[cli]==<latest-version>
azure-ai-formrecognizer
azure-search-documents
azure-ai-embeddings
python-dotenv
